FBX export. to_ktime ticks was ok.

+++ code/FBXExporter.cpp
@@ -980,9 +980,9 @@ aiMatrix4x4 get_world_transform(const aiNode* node, const aiScene* scene)
 
 int64_t to_ktime(double ticks, const aiAnimation* anim) {
     if (anim->mTicksPerSecond <= 0) {
-        return static_cast<int64_t>(ticks * FBX::SECOND);
+        return static_cast<int64_t>(ticks) * FBX::SECOND;
     }
-    return (static_cast<int64_t>(ticks / anim->mTicksPerSecond * FBX::SECOND));
+    return (static_cast<int64_t>(ticks) / static_cast<int64_t>(anim->mTicksPerSecond)) * FBX::SECOND;
 }
 
 int64_t to_ktime(double time) {
